@@ -15,7 +15,8 @@
 
 from __future__ import annotations
 
-from typing import Any, Type
+import abc
+from typing import Any
 
 from neo4j_graphrag.embeddings.base import Embedder
 
@@ -25,26 +26,29 @@ except ImportError:
     openai = None  # type: ignore
 
 
-class OpenAIEmbeddings(Embedder):
-    """
-    OpenAI embeddings class.
-    This class uses the OpenAI python client to generate embeddings for text data.
-
-    Args:
-        model (str): The name of the OpenAI embedding model to use. Defaults to "text-embedding-ada-002".
-    """
-
-    client_class: Type[openai.OpenAI] = openai.OpenAI
-
+class BaseOpenAIEmbeddings(Embedder, abc.ABC):
     def __init__(self, model: str = "text-embedding-ada-002", **kwargs: Any) -> None:
         if openai is None:
             raise ImportError(
                 "Could not import openai python client. "
                 "Please install it with `pip install openai`."
             )
-
-        self.openai_model = self.client_class(**kwargs)
         self.model = model
+
+
+class OpenAIEmbeddings(BaseOpenAIEmbeddings):
+    """
+    OpenAI embeddings class.
+    This class uses the OpenAI python client to generate embeddings for text data.
+
+    Args:
+        model (str): The name of the OpenAI embedding model to use. Defaults to "text-embedding-ada-002".
+        kwargs: All other parameters will be passed to the openai.OpenAI init.
+    """
+
+    def __init__(self, model: str = "text-embedding-ada-002", **kwargs: Any) -> None:
+        super().__init__(model, **kwargs)
+        self.openai_client = openai.OpenAI(**kwargs)
 
     def embed_query(self, text: str, **kwargs: Any) -> list[float]:
         """
@@ -54,11 +58,13 @@ class OpenAIEmbeddings(Embedder):
             text (str): The text to generate an embedding for.
             **kwargs (Any): Additional arguments to pass to the OpenAI embedding generation function.
         """
-        response = self.openai_model.embeddings.create(
+        response = self.openai_client.embeddings.create(
             input=text, model=self.model, **kwargs
         )
         return response.data[0].embedding
 
 
 class AzureOpenAIEmbeddings(OpenAIEmbeddings):
-    client_class: Type[openai.OpenAI] = openai.AzureOpenAI
+    def __init__(self, model: str = "text-embedding-ada-002", **kwargs: Any) -> None:
+        super().__init__(model, **kwargs)
+        self.openai_client = openai.AzureOpenAI(**kwargs)
